Stop truncating the face index to uint8_t in three internal tools
prc_api_get_graphics_primitive takes a uint32_t face_index, and has since
the initial public release -- there was never a narrower signature these
casts were written against. Casting to uint8_t silently wraps the index
every 256 faces, so on any uncompressed tessellation with more faces than
that, the tools query the wrong face's primitives, repeatedly, while
offsetting the indices by the correct face's vertex offset.

Found while resolving a triangle-count discrepancy on a 3,739-face
uncompressed file, where a throwaway measuring tool that had copied the
cast from reauthor_tess_pdf.c reported 396,890 triangles. With the cast
removed it reports 231,325 -- matching an independent non-nanoPRC reader
exactly, and matching nano_prc_stl_export's 230,973 written triangles once
its 352 degenerate-triangle rejections are accounted for.

Scope is narrow but the failure is silent, which is what makes it worth
fixing: the wrong count is plausible rather than obviously broken, and
these are diagnostic tools whose output gets quoted in investigations.
detect_tess_spikes.c in particular reports per-triangle geometry that is
read back as evidence.

Only uncompressed (PRC_TESS_3D) tessellations with more than 255 faces are
affected. On the compressed path face_limit is 1, so the cast never bit,
and f is uint32_t in every one of these call sites already -- the cast was
narrowing a value that was correct to begin with.

No behaviour change for any file with 256 faces or fewer.

Verified: all three targets rebuild clean; tests/internal/check_c99.py
reports all 96 first-party C sources parse clean at -std=c99;
detect_tess_spikes now walks the full triangle range on the 3,739-face
file (indices past 215,000, previously unreachable).

+++ tests/internal/reauthor_tess_pdf.c
@@ -215,7 +215,7 @@ int main(int argc, char **argv)
             {
                 prc_api_graphic_primitive prim;
                 size_t idx;
-                code = prc_api_get_graphics_primitive(ctx, data, &tess, (uint8_t)f, p, &prim);
+                code = prc_api_get_graphics_primitive(ctx, data, &tess, f, p, &prim);
                 if (code < 0) continue;
 
 #define APPEND_TRI(a,b,c) do { \
